move Dashboard into its own component
