+++ frontend/src/components/TagsCheckboxes/index.js
@@ -1,3 +1,3 @@
-import TagsCheckBoxes from "./TagsCheckboxes";
+import TagsCheckBoxes from './TagsCheckboxes'
 
 export default TagsCheckBoxes
