+++ frontend/src/context/context.js
@@ -1,4 +1,4 @@
-import { useState, useContext, createContext } from 'react'
+import { useState, createContext } from 'react'
 import API from '../config/API'
 
 const useAPI = () => {
